ProjectPRT-BE/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.109.0
google-api-core==2.28.1
google-auth==2.32.0
google-cloud-core==2.5.0
google-cloud-storage==2.14.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
proto-plus==1.27.0
protobuf==6.33.2
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
PyJWT==2.9.0
python-dotenv==1.2.1
python-jose==3.5.0
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.35.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.27.0
email-validator==2.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
google-cloud-aiplatform>=1.38.0